+++ source/locale.js
@@ -3,33 +3,28 @@
 // based on the list of preferred `locales` supplied by the user.
 //
 // @param {string[]} locales - the list of preferable locales (in [IETF format](https://en.wikipedia.org/wiki/IETF_language_tag)).
-// @param {Object} availableLocales - a map of available locales.
+// @param {Function} isLocaleDataAvailable - tests if a locale is available.
 //
 // @returns {string} The most suitable locale
 //
 // @example
 // // Returns 'en'
-// chooseLocale(['en-US'], undefined, { 'ru', 'en' })
+// chooseLocale(['en-US'], undefined, (locale) => locale === 'ru' || locale === 'en')
 //
-export default function chooseLocale(locales, availableLocales)
+export default function chooseLocale(locales, isLocaleDataAvailable)
 {
 	// This is not an intelligent algorythm,
 	// but it will do for this library's case.
 	// `sr-Cyrl-BA` -> `sr-Cyrl` -> `sr`.
-	for (let locale of locales)
-	{
-		if (availableLocales[locale])
-		{
+	for (let locale of locales) {
+		if (isLocaleDataAvailable(locale)) {
 			return locale
 		}
-
 		const parts = locale.split('-')
-		while (parts.length > 1)
-		{
+		while (parts.length > 1) {
 			parts.pop()
 			locale = parts.join('-')
-			if (availableLocales[locale])
-			{
+			if (isLocaleDataAvailable(locale)) {
 				return locale
 			}
 		}
@@ -44,11 +39,9 @@ export default function chooseLocale(locales, availableLocales)
  * @param  {(string|string[])} locales
  * @return {?string} The first locale that can be used.
  */
-export function intlDateTimeFormatSupportedLocale(locales)
-{
+export function intlDateTimeFormatSupportedLocale(locales) {
 	/* istanbul ignore else */
-	if (intlDateTimeFormatSupported())
-	{
+	if (intlDateTimeFormatSupported()) {
 		return Intl.DateTimeFormat.supportedLocalesOf(locales)[0]
 	}
 }
@@ -56,12 +49,10 @@ export function intlDateTimeFormatSupportedLocale(locales)
  * Whether can use `Intl.DateTimeFormat`.
  * @return {boolean}
  */
-export function intlDateTimeFormatSupported()
-{
+export function intlDateTimeFormatSupported() {
 	// Babel transforms `typeof` into some "branches"
 	// so istanbul will show this as "branch not covered".
 	/* istanbul ignore next */
 	const isIntlAvailable = typeof Intl === 'object'
-
 	return isIntlAvailable && typeof Intl.DateTimeFormat === 'function'
 }